third_party/WebKit/LayoutTests/http/tests/feature-policy-experimental-features/vibrate-allowed-by-container-policy.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/helper.js"></script>
<iframe allow="vibrate"></iframe>
<script>
var srcs = [
  "resources/feature-policy-vibrate.html",
  "http://localhost:8000/feature-policy-experimental-features/resources/feature-policy-vibrate.html"];

window.onload = function () {
  var iframe = document.querySelector('iframe');
  iframe.addEventListener('load', sendClick);
  function loadFrame(src) {
    promise_test(function() {
      iframe.src = src;
      return new Promise(function(resolve, reject) {
        window.addEventListener('message', function(e) {
          if (e.data.type === 'result') {
            resolve(e.data);
          }
        });
      }).then(function(data) {
        assert_true(data.enabled, 'navigator.vibrate():');
      });
    }, 'Navigator.vibrate enabled by container policy on URL: ' + src);
  }
  for (var src of srcs) {
    loadFrame(src);
  }
}
</script>
